c program

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -11,7 +11,7 @@ void rev_string(char *s)
 
 	char tmp;
 
-	int length = strlen(s);
+	int length = (int)strlen(s);
 
 	for (i = 0; i < (length / 2); i++)
 	{
